pyproject.toml
[project]
name = "textbook-marketplace"
version = "0.1.0"
description = ""
authors = [
    { name = "mrezvova", email = "rezvmaria@gmail.com" },
]
readme = "README.md"
requires-python = "~=3.12"

dependencies = [
    "django==5.1.7",
    "psycopg2-binary==2.9.9",
    "django-cors-headers==4.4.0",
    "django-filter==24.3",
    "djangorestframework==3.15.2",
    "pillow==11.0.0",
    "django-allauth==65.2.0",
    "djangorestframework-simplejwt==5.3.1",
    "rest-framework-simplejwt==0.0.2",
    "wheel==0.45.1",
    "faker==35.0.0",
    "django-versatileimagefield==3.1",
    "python-magic-bin==0.4.14 ; sys_platform == 'win32'",
    "python-decouple==3.8",
    "channels[daphne]==4.2.0",
    "channels-redis==4.2.1",
    "django-channels-jwt-auth-middleware==1.0.0",
    "gunicorn==21.2.0",
    "django-ratelimit==4.1.0",
    "bleach==6.1.0",
    "drf-spectacular==0.27.1",
    "structlog==24.1.0"

]

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "pytest-django==4.8",
    "black==23.3",
    "flake8==6.0",
    "isort==5.12",
    "whitenoise==6.9.0",
    "pytest-asyncio==0.26.0",
    "pytest-xdist[psutil]==3.6.1",
    "pytest-cov==6.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
